fix conflict in ServerPreferencesFragment
