--- collect_app/src/main/java/org/odk/collect/android/preferences/CustomEditTextPreference.java
@@ -1,13 +1,0 @@
-package org.odk.collect.android.preferences;
-
-import android.content.Context;
-import android.util.AttributeSet;
-
-import androidx.preference.EditTextPreference;
-
-public class CustomEditTextPreference extends EditTextPreference {
-
-    public CustomEditTextPreference(Context context, AttributeSet attrs) {
-        super(context, attrs);
-    }
-}
