+++ collect_app/src/main/res/xml/google_preferences.xml
@@ -6,7 +6,7 @@
         android:title="@string/selected_google_account_text"
         app:iconSpaceReserved="false" />
 
-    <org.odk.collect.android.preferences.CustomEditTextPreference
+    <EditTextPreference
         android:dialogTitle="@string/google_sheets_url"
         android:inputType="textNoSuggestions"
         android:key="google_sheets_url"
